split dependencies.py (#1553)
